requirements.txt
aiohttp==3.8.4
aiomoex==2.0.0
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cvxpy==1.3.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
ecos==2.0.12
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.56.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.1
osqp==0.6.3
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.0
plotly==5.15.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
qdldl==0.1.7.post0
referencing==0.29.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tinkoff==0.1.1
tinkoff-investments==0.2.0b58
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
yarl==1.9.2
